carceral/requirements.txt
gwpy
matplotlib
requests
scipy
tabula-py
tqdm
